God commit: - Install HWIOAuthBundle - Rename BankIdService to BankIdResourceOwner according to HWI interfaces - Remove custom ugly login method
