+++ app/config/routing.yml
@@ -2,6 +2,16 @@ app:
     resource: "@AppBundle/Controller/"
     type:     annotation
 
+hwi_oauth_redirect:
+    resource: "@HWIOAuthBundle/Resources/config/routing/redirect.xml"
+    prefix:   /connect
+
+hwi_oauth_login:
+    resource: "@HWIOAuthBundle/Resources/config/routing/login.xml"
+    prefix:   /login
+
+bank_id_login:
+    path: /login/check-bank-id
 
 admin_login_check:
     path: /admin/login_check
