+++ app/config/security.yml
@@ -1,5 +1,14 @@
 # To get started with security, check out the documentation:
 # http://symfony.com/doc/current/book/security.html
+hwi_oauth:
+    firewall_name: main
+    resource_owners: ~
+    # uncomment next lines when BankId will commited in HWIOauthBundle
+#        bank_id:
+#            type: bank_id
+#            client_id: %client_id%
+#            client_secret: %client_secret%
+
 security:
     encoders:
         AppBundle\Entity\Admin:
@@ -45,6 +54,18 @@ security:
         main:
             pattern: ^/
             anonymous: ~
+            oauth:
+                login_path:        /login
+                use_forward:       false
+                failure_path:      /login
+                resource_owners:
+                    bank_id: "/login/check-bank-id"
+
+                oauth_user_provider:
+                    orm:
+                        class: AppBundle\Entity\User
+                        properties: {'bank_id': 'clid'}
+                        manager_name: default
             form_login:
                 provider: user
                 login_path: login
