+++ src/AppBundle/AppBundle.php
@@ -2,8 +2,16 @@
 
 namespace AppBundle;
 
+use AppBundle\DependencyInjection\Compiler\SetPossibleResourceOwnersCompilerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class AppBundle extends Bundle
 {
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new SetPossibleResourceOwnersCompilerPass());
+    }
 }
