+++ src/AppBundle/DependencyInjection/Compiler/SetPossibleResourceOwnersCompilerPass.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace AppBundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+class SetPossibleResourceOwnersCompilerPass implements CompilerPassInterface
+{
+    /**
+     * You can modify the container here before it is dumped to PHP code.
+     *
+     * @param ContainerBuilder $container
+     *
+     * @api
+     */
+    public function process(ContainerBuilder $container)
+    {
+        $container->getDefinition('hwi_oauth.resource_ownermap.main')->replaceArgument(1, ['bank_id']);
+    }
+}
