+++ src/AppBundle/Resources/config/services.yml
@@ -2,23 +2,26 @@ parameters:
 #    app.example.class: AppBundle\Example
 
 services:
-    app.security.bank_id:
-        class: AppBundle\Security\BankIdService
+    hwi_oauth.resource_owner.bank_id:
+        class: AppBundle\Security\BankIdResourceOwner
         arguments:
-            - %client_id%
-            - %client_secret%
-            - @router
-            - @monolog.logger
-            - @validator
-            - @jms_serializer.serializer
+            - @buzz.client
+            - @security.http_utils
+            - client_id: %client_id%
+              client_secret: %client_secret%
+              authorization_url: https://bankid.privatbank.ua/DataAccessService/das/authorize
+              access_token_url: https://bankid.privatbank.ua/DataAccessService/oauth/token
+              infos_url: https://bankid.privatbank.ua/ResourceService/checked/data
+            - "bank_id"
+            - @hwi_oauth.storage.session
     app.twig_extension:
         class: AppBundle\Twig\AppExtension
         public: false
         arguments:
-            - @app.security.bank_id
+            - @hwi_oauth.resource_owner.bank_id
+            - @router
         tags:
             - { name: twig.extension }
-
     app.user.manager:
         class:  AppBundle\Helper\UserManager
         arguments: [@templating, @doctrine.orm.entity_manager ]
